refactor: split monolithic server.js into route modules and services (#217)
