+++ config/firebase.config.js
@@ -1,6 +1,19 @@
 const admin = require('firebase-admin');
 require('dotenv').config();
 
+/** @type {admin.firestore.Firestore | null} */
+let db = null;
+
+/** @type {admin.auth.Auth | null} */
+let auth = null;
+
+/** @type {{ enabled: boolean, mode: string, reason: string }} */
+const firebaseState = {
+  enabled: false,
+  mode: 'memory',
+  reason: 'Firebase not initialized'
+};
+
 let isInitialized = false;
 
 /**
@@ -20,11 +33,18 @@ function initializeFirebase() {
         privateKey: process.env.FIREBASE_PRIVATE_KEY?.replace(/\\n/g, '\n')
       })
     });
-    console.log('✅ Firebase Admin initialized');
+
+    db = admin.firestore();
+    auth = admin.auth();
+    firebaseState.enabled = true;
+    firebaseState.mode = 'firestore';
+    firebaseState.reason = 'Firebase connected successfully';
     isInitialized = true;
+
+    console.log('✅ Firebase Admin initialized');
     return true;
   } catch (error) {
-    console.log('⚠️  Firebase Admin not configured. Using in-memory storage.');
+    console.log('⚠️ Firebase Admin not configured. Using in-memory storage.');
     console.log('   See docs/FIREBASE_SETUP.md for setup instructions.');
     return false;
   }
@@ -32,18 +52,26 @@ function initializeFirebase() {
 
 /**
  * Get Firestore database instance
- * @returns {admin.firestore.Firestore}
+ * @returns {admin.firestore.Firestore | null}
  */
 function getDatabase() {
-  return admin.firestore();
+  return db;
 }
 
 /**
  * Get Firebase Auth instance
- * @returns {admin.auth.Auth}
+ * @returns {admin.auth.Auth | null}
  */
 function getAuth() {
-  return admin.auth();
+  return auth;
+}
+
+/**
+ * Get current Firebase state
+ * @returns {{ enabled: boolean, mode: string, reason: string }}
+ */
+function getFirebaseState() {
+  return firebaseState;
 }
 
 /**
@@ -52,13 +80,15 @@ function getAuth() {
 const COLLECTIONS = {
   LINKS: 'links',
   ANALYTICS: 'analytics',
-  USERS: 'users'
+  USERS: 'users',
+  BIO_LINKS: 'bioLinks'
 };
 
 module.exports = {
   initializeFirebase,
   getDatabase,
   getAuth,
+  getFirebaseState,
   admin,
   COLLECTIONS
 };
